+++ nf_core/modules/install.py
@@ -1,20 +1,16 @@
 import logging
 import os
 
-import questionary
-
 import nf_core.components.components_install
 import nf_core.modules.modules_utils
 import nf_core.utils
+from nf_core.components.components_command import ComponentCommand
 from nf_core.modules.modules_json import ModulesJson
-
-from .modules_command import ModuleCommand
-from .modules_repo import NF_CORE_MODULES_NAME
 
 log = logging.getLogger(__name__)
 
 
-class ModuleInstall(ModuleCommand):
+class ModuleInstall(ComponentCommand):
     def __init__(
         self,
         pipeline_dir,
@@ -24,11 +20,16 @@ class ModuleInstall(ModuleCommand):
         remote_url=None,
         branch=None,
         no_pull=False,
+        installed_by=False,
     ):
-        super().__init__(pipeline_dir, remote_url, branch, no_pull)
+        super().__init__("modules", pipeline_dir, remote_url, branch, no_pull)
         self.force = force
         self.prompt = prompt
         self.sha = sha
+        if installed_by:
+            self.installed_by = installed_by
+        else:
+            self.installed_by = self.component_type
 
     def install(self, module, silent=False):
         if self.repo_type == "modules":
@@ -69,8 +70,13 @@ class ModuleInstall(ModuleCommand):
 
         # Check that the module is not already installed
         if not nf_core.components.components_install.check_component_installed(
-            self.component_type, module, current_version, module_dir, self.modules_repo, self.force
+            self.component_type, module, current_version, module_dir, self.modules_repo, self.force, self.prompt
         ):
+            log.debug(
+                f"Module is already installed and force is not set.\nAdding the new installation source {self.installed_by} for module {module} to 'modules.json' without installing the module."
+            )
+            modules_json.load()
+            modules_json.update(self.modules_repo, module, current_version, self.installed_by)
             return False
 
         version = nf_core.components.components_install.get_version(
@@ -80,10 +86,11 @@ class ModuleInstall(ModuleCommand):
             return False
 
         # Remove module if force is set
+        install_track = None
         if self.force:
             log.info(f"Removing installed version of '{self.modules_repo.repo_path}/{module}'")
             self.clear_component_dir(module, module_dir)
-            nf_core.components.components_install.clean_modules_json(
+            install_track = nf_core.components.components_install.clean_modules_json(
                 module, self.component_type, self.modules_repo, modules_json
             )
 
@@ -103,5 +110,5 @@ class ModuleInstall(ModuleCommand):
 
         # Update module.json with newly installed module
         modules_json.load()
-        modules_json.update(self.modules_repo, module, version)
+        modules_json.update(self.modules_repo, module, version, self.installed_by, install_track)
         return True