+++ nf_core/modules/patch.py
@@ -7,17 +7,17 @@ from pathlib import Path
 import questionary
 
 import nf_core.utils
+from nf_core.components.components_command import ComponentCommand
 
-from .modules_command import ModuleCommand
 from .modules_differ import ModulesDiffer
 from .modules_json import ModulesJson
 
 log = logging.getLogger(__name__)
 
 
-class ModulePatch(ModuleCommand):
+class ModulePatch(ComponentCommand):
     def __init__(self, dir, remote_url=None, branch=None, no_pull=False):
-        super().__init__(dir, remote_url, branch, no_pull)
+        super().__init__("modules", dir, remote_url, branch, no_pull)
 
         self.modules_json = ModulesJson(dir)
 
@@ -62,7 +62,9 @@ class ModulePatch(ModuleCommand):
                 f"The '{module_fullname}' module does not have a valid version in the 'modules.json' file. Cannot compute patch"
             )
         # Get the module branch and reset it in the ModulesRepo object
-        module_branch = self.modules_json.get_module_branch(module, self.modules_repo.remote_url, module_dir)
+        module_branch = self.modules_json.get_component_branch(
+            self.component_type, module, self.modules_repo.remote_url, module_dir
+        )
         if module_branch != self.modules_repo.branch:
             self.modules_repo.setup_branch(module_branch)